pyproject.toml
[tool.poetry]
name = "ticketmyra"
version = "0.1.0"
description = "a ticketmaster api for filter what you need"
authors = ["myticketra"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.2"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
myst-nb = "^0.17.1"
sphinx-autoapi = "^2.0.0"
sphinx-rtd-theme = "^1.1.1"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
